requirements/base.txt
# Core Python
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
# Web Framework
Flask==2.3.2
Flask-RESTX==1.1.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
# Database
SQLAlchemy==2.0.17
psycopg2-binary==2.9.7
alembic==1.12.0
# Task Queue
celery==5.3.1
redis==5.0.1
# HTTP Client
requests==2.31.0
httpx==0.24.1
# Data Processing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0
# Medical Imaging
monai==1.2.0
torch==2.0.1
torchvision==0.15.2
pydicom==2.3.1
SimpleITK==2.2.1
nibabel==5.1.0
opencv-python-headless==4.8.0.74
# Federated Learning
flwr==1.4.0
syft==0.8.2
opacus==1.4.0
# ML Operations
mlflow==2.5.0
boto3==1.28.45
prometheus-client==0.17.1
# Utilities
click==8.1.4
tqdm==4.65.0
pyyaml==6.0
colorlog==6.7.0
# Security
cryptography==41.0.2
PyJWT==2.8.0
bcrypt==4.0.1
